separate out WorkspaceItemsTable components and fix filtering issue
