+++ context/app/static/js/components/workspaces/NewWorkspaceDialog/NewWorkspaceDialogFromSelections.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import Box from '@mui/material/Box';
 import MenuItem from '@mui/material/MenuItem';
-import SvgIcon from '@mui/material/SvgIcon';
 import { useEventCallback } from '@mui/material/utils';
+import ListItemIcon from '@mui/material/ListItemIcon';
 
 import NewWorkspaceDialog from 'js/components/workspaces/NewWorkspaceDialog';
 import ErrorOrWarningMessages from 'js/shared-styles/alerts/ErrorOrWarningMessages';
-import WorkspacesIcon from 'assets/svg/workspaces.svg';
 import { useSelectableTableStore } from 'js/shared-styles/tables/SelectableTableProvider';
+import { WorkspacesIcon } from 'js/shared-styles/icons';
 import { trackEvent } from 'js/helpers/trackers';
 import { WorkspacesEventCategories } from 'js/components/workspaces/types';
 import { useCreateWorkspaceDatasets, useCreateWorkspaceForm } from './useCreateWorkspaceForm';
@@ -42,7 +42,9 @@ function NewWorkspaceDialogFromSelections() {
   return (
     <>
       <MenuItem onClick={handleClick}>
-        <SvgIcon component={WorkspacesIcon} sx={{ mr: 1, fontSize: '1.25rem' }} color="primary" />
+        <ListItemIcon>
+          <WorkspacesIcon fontSize="1.25rem" color="primary" />
+        </ListItemIcon>
         Create New Workspace
       </MenuItem>
       <NewWorkspaceDialog
